+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Unreleased
 0.11.0 - 2021/01/23
 -----------
 ### Added
+- add linearoperator util: is_XXX https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/227
 - LinerOperator https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/210
 - LOBPCG/DC works on GPU, too https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/220
 - add vector is_same_size, is_same_structure https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/224
@@ -39,6 +40,7 @@ Unreleased
 - add VML max/min https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/215
 
 ### Changed
+- organize linearoperator utils https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/227
 - compute_hash in CRS convert and constructor https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/221
 - make internal syev()/sygv() interface to Fortran95-like https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/219
 - Summarize CI stage public function https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/218